PCM-1527 PCM-1516 pr feedback
